# Avoid deadlock between Close when offset-modifying operation is still pending