--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -12,10 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
-        go: ['1.23', '1.22']
-        exclude:
-          - os: macos-latest
-            go: '1.22'
+        go: ['1.23']
 
     steps:
       - uses: actions/checkout@v4