--- a/localfs/statvfs/statvfs_stubs.go
+++ b/localfs/statvfs/statvfs_stubs.go
@@ -10,10 +10,10 @@ import (
 	"github.com/pkg/sftp/v2/encoding/ssh/filexfer/openssh"
 )
 
-// StatVFS stubs the OpenSSH StatVFS with an sshfx.StatusOPUnsupported Status.
+// StatVFS stubs the OpenSSH StatVFS with an sshfx.StatusOpUnsupported Status.
 func StatVFS(name string) (*openssh.StatVFSExtendedReplyPacket, error) {
 	return nil, &sshfx.StatusPacket{
-		StatusCode:   sshfx.StatusOPUnsupported,
+		StatusCode:   sshfx.StatusOpUnsupported,
 		ErrorMessage: "not supported by " + runtime.GOOS,
 	}
 }